requirements.txt
selenium==4.19.0
pytest==8.1.1
pyautogui==0.9.54
randominfo==2.0.2
allure-pytest==2.13.5